refactor: api module DTOs
