+++ src/modules/api/pipes/QueryValidator.pipe.ts
@@ -3,7 +3,7 @@ import SchemaValidator from '@common/utils/validators/SchemaValidator.validator'
 import Exceptions from '@core/errors/Exceptions';
 import listQuerySchema from '@api/schemas/listQuery.schema';
 import { ListQueryInterface } from '@shared/internal/interfaces/listPaginationInterface';
-import { ListQueryInputDto } from './dto/QueryInput.dto';
+import { ListQueryInputDto } from '../dto/QueryInput.dto';
 import { generateLogger } from '@core/logging/logger';
